refactor: Decompose monolithic doctor.rs into focused submodules
Split the 1,624-line doctor.rs into coherent units by responsibility:
- types.rs: DoctorRequest, DoctorMode, Problem, and diagnostic records
- inspect.rs: Environment, config, repo, and hook inspection logic
- render.rs: Text/JSON output formatting and human-readable diagnostics
- fixes.rs: Repair-mode hook installation and DB bootstrap routines

This reduces per-file complexity while maintaining the stable doctor
command surface, JSON/text output contracts, and problem taxonomy.
