+++ context/architecture.md
@@ -106,7 +106,7 @@ The repository includes a new placeholder Rust binary crate at `cli/`.
 - `cli/src/test_support.rs` provides a shared test-only temp-directory helper (`TestTempDir`) used by service tests that need filesystem fixtures.
 - `cli/src/services/setup.rs` defines the setup command contract (`SetupMode`, `SetupTarget`, `SetupRequest`, CLI flag parser/validator), an `inquire`-backed interactive target prompter (`InquireSetupTargetPrompter`), setup dispatch outcomes (proceed/cancelled), compile-time embedded asset access (`EmbeddedAsset`, target-scoped iterators, required-hook asset iterators/lookups) generated by `cli/build.rs` from the ephemeral crate-local `cli/assets/generated/config/{opencode,claude}/**` mirror plus `cli/assets/hooks/**`, a target-scoped install engine/orchestrator that stages embedded files and uses a unified remove-and-replace policy (removing existing targets before swapping staged content, with deterministic recovery guidance on swap failure and no backup artifact creation), and formats deterministic completion messaging, plus required-hook install orchestration (`install_required_git_hooks`) that follows the same remove-and-replace policy (removing existing hooks before swapping staged content, with deterministic recovery guidance on swap failure), along with command-surface setup request resolution helpers (`run_setup_hooks`, `resolve_setup_request`) used by hooks-only and composable target+hooks setup invocations with deterministic option compatibility validation, canonicalized/validated repo targeting, write-permission probes, and stable section-ordered setup/hook outcome messaging.
 - `cli/src/services/security.rs` provides shared security utilities for deterministic secret redaction (`redact_sensitive_text`) and directory write-permission probes (`ensure_directory_is_writable`) used by app/setup/observability surfaces.
-- `cli/src/services/doctor.rs` now defines the current doctor request/report surface: explicit `DoctorMode` (`diagnose` vs `fix`), stable text/JSON problem records with category/severity/fixability/remediation fields, deterministic fix-result reporting in fix mode, current global operator-environment checks (shared state-root resolution, shared global-config-path resolution, repo-local config validation through the generated `sce/config.json` schema contract, local DB path/health, and DB-parent readiness barriers), repo/hook-integrity diagnostics that distinguish git-unavailable vs non-repo vs bare-repo states, an empty default repo-database section reflecting the current state where no repo-owned SCE database exists, repair-mode reuse of `cli/src/services/setup.rs::install_required_git_hooks` for missing/stale/non-executable required hooks and missing hooks directories, and a bounded doctor-owned repair routine that bootstraps the canonical SCE-owned local DB parent directory only when the resolved path matches the expected owned location.
+- `cli/src/services/doctor/mod.rs` now owns the current doctor request/report surface while focused submodules (`doctor/inspect.rs`, `doctor/render.rs`, `doctor/fixes.rs`, `doctor/types.rs`) split diagnosis, rendering, fix execution, and doctor-owned domain types into smaller seams. Together they preserve explicit `DoctorMode` (`diagnose` vs `fix`), stable text/JSON problem records with category/severity/fixability/remediation fields, deterministic fix-result reporting in fix mode, current global operator-environment checks (shared state-root resolution, shared global-config-path resolution, repo-local config validation through the generated `sce/config.json` schema contract, local DB path/health, and DB-parent readiness barriers), repo/hook-integrity diagnostics that distinguish git-unavailable vs non-repo vs bare-repo states, an empty default repo-database section reflecting the current state where no repo-owned SCE database exists, repair-mode reuse of `cli/src/services/setup.rs::install_required_git_hooks` for missing/stale/non-executable required hooks and missing hooks directories, and a bounded doctor-owned repair routine that bootstraps the canonical SCE-owned local DB parent directory only when the resolved path matches the expected owned location.
 - `cli/src/services/version.rs` defines the version command parser/rendering contract (`parse_version_request`, `render_version`) with deterministic text output and stable JSON runtime-identification fields.
 - `cli/src/services/completion.rs` defines completion parser/rendering contract (`parse_completion_request`, `render_completion`) with deterministic Bash/Zsh/Fish script output aligned to current parser-valid command/flag surfaces.
 - `cli/src/services/hooks.rs` defines the current minimal local hook runtime parsing/dispatch (`HookSubcommand`, `run_hooks_subcommand`) plus a commit-msg co-author policy seam (`apply_commit_msg_coauthor_policy`) that injects one canonical SCE trailer only when the disabled-default attribution-hooks config/env control is enabled and `SCE_DISABLED` is false. The other hook entrypoints (`pre-commit`, `post-commit`, `post-rewrite`) are deterministic no-op surfaces in the current attribution-only baseline.
